+++ CodeCpp/Score.h
@@ -17,17 +17,17 @@
 class Score {
     private :
         std::vector<int> my_success; //Notre tableau contenant le score de tout nos level
-    
+
     public :
         Score();
         int getCurrent() const; //Connaitre le score courant
-        long getCurrentStep() const; //Connaitre le niveau
+        int getCurrentStep() const; //Connaitre le niveau
         int getGlobale() const ; //Connaitre le score global
         void addSuccess(); //Ajoute une case
         void resetScore(); //Remet à zéro le score courant
         void addPoints(const int &i); //Ajoute des points à la case courrante
         void showTmpScore( int langue ) const;
-    
+
 };
 
 #endif
